requirements.txt
# ── Core ──────────────────────────────────────────────────────
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# ── Database ──────────────────────────────────────────────────
psycopg2-binary==2.9.9        # PostgreSQL driver
dj-database-url==2.1.0        # Parse DATABASE_URL env var

# ── Auth & Security ───────────────────────────────────────────
django-allauth==0.57.0        # Social auth (Google, Apple)
PyJWT==2.8.0                  # JWT tokens
cryptography==41.0.7          # For token signing
django-cors-headers==4.3.1    # CORS for React Native / web

# ── Environment ───────────────────────────────────────────────
python-dotenv==1.0.0          # .env file loading
django-environ==0.11.2        # Django environ integration

# ── Validation ────────────────────────────────────────────────
django-filter==23.5           # Query filtering
djangorestframework-camel-case==1.4.2  # camelCase ↔ snake_case

# ── Payments ──────────────────────────────────────────────────
razorpay==1.4.1               # Razorpay payment gateway

# ── Utilities ─────────────────────────────────────────────────
Pillow==10.2.0                # Image handling
django-storages==1.14.2       # S3/GCS storage
boto3==1.34.11                # AWS S3 (optional)
celery==5.3.6                 # Background tasks
redis==5.0.1                  # Celery broker / cache
django-redis==5.4.0           # Django cache with Redis

# ── Email ─────────────────────────────────────────────────────
django-anymail==10.2          # Email providers (SendGrid, etc.)

# ── Monitoring ────────────────────────────────────────────────
sentry-sdk==1.39.1            # Error tracking

# ── Development only ──────────────────────────────────────────
django-debug-toolbar==4.2.0   # Debug toolbar (dev only)
factory-boy==3.3.0            # Test factories
faker==22.0.0                 # Fake data generation
pytest-django==4.7.0          # Pytest for Django
coverage==7.4.0               # Test coverage
